Finish off FLAC implementation
Obviously it doesn't work in any way, shape, or form; but the rest should just be bugfix.

+++ src/com/stefankopieczek/audinance/audiosources/SimpleEncodedSource.java
@@ -25,10 +25,10 @@ public class SimpleEncodedSource extends EncodedSource
 	
 	@Override
 	public byte getByte(int index) throws NoMoreDataException
-	{		
+	{
 		if (index < mData.length)
 		{	
-			return mData[index];
+			return mData[(int)index];
 		}
 		else
 		{
