+++ src/com/stefankopieczek/audinance/formats/wav/WavDecoder.java
@@ -29,11 +29,11 @@ public class WavDecoder
 	}
 	
 	/**
-	 * Gets an array of bytes of the given length from the data sourcr, starting
+	 * Gets an array of bytes of the given length from the data source, starting
 	 * at the given index.
 	 * TODO: Why isn't this a method of EncodedSource?
 	 *
-	 * @param start The byte index tonstart reading from in the encoded source.
+	 * @param start The byte index to start reading from in the encoded source.
 	 * @param length The number of bytes to read.
 	 * @return 'length' bytes from mWavSource, starting at index 'start'.
 	 */
@@ -89,18 +89,17 @@ public class WavDecoder
 			final int finalChannel = channel;
 			channels[channel] = new DecodedSource()
 			{
-				public double getSample(int idx) 
+				public double getSample(int idx)
 					throws InvalidWavDataException, NoMoreDataException
 				{
 					// The index of the individual bit in the wav data that begins the frame
 					// containing the desired sample.
-					int frameStartIdx = fmtChunk.getBitsPerSample() * 
+					int frameStartIdx = fmtChunk.getBitsPerSample() *
 							            idx * 
 							            fmtChunk.getNumChannels();
 										
 					// The number of bits at the start of the frame before the desired sample.
-					int offsetToSample = finalChannel * 
-							                          fmtChunk.getBitsPerSample();
+					int offsetToSample = finalChannel * fmtChunk.getBitsPerSample();
 													  
 					// Get the sample by specifying the index as a byte.
 					return dataChunk.getSample((frameStartIdx + offsetToSample) / 8);
@@ -371,11 +370,11 @@ public class WavDecoder
 			return CHUNK_SIZE_IDX_OFFSET;
 		}
 		
-		public double getSample(int byteIdx) 
+		public double getSample(int byteIdx)
 				throws InvalidWavDataException, NoMoreDataException
 		{
 			double result;						
-			int endOfSample = (int)(byteIdx + Math.ceil(mBitsPerSample / 8.0)); 
+			long endOfSample = (long)(byteIdx + Math.ceil(mBitsPerSample / 8.0));
 			if (endOfSample >= getLength())
 			{
 				throw new NoMoreDataException();
